Add struct capturing join plan expansion
